Convert registry image Docker Hub test to YAML
