--- examples/test-registry-image-dockerhub/ts/package.json
@@ -1,10 +1,0 @@
-{
-    "name": "test-registry-image-dockerhub-ts",
-    "devDependencies": {
-        "@types/node": "^18"
-    },
-    "dependencies": {
-        "@pulumi/pulumi": "^3",
-        "@pulumi/docker": "^4"
-    }
-}
